Consolidate Polaris and Shadcn storybook together with a UI lib selector to make it easier to maintain feature parity
