+++ packages/react/spec/auto/shadcn-defaults/inputs/ShadcnAutoEnumInput.stories.jsx
@@ -1,11 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 import { Provider } from "../../../../src/GadgetProvider.tsx";
+import { makeAutocomponents } from "../../../../src/auto/shadcn/index.ts";
+import { makeShadcnAutoEnumInput } from "../../../../src/auto/shadcn/inputs/ShadcnAutoEnumInput.tsx";
 import { FormProvider, useForm } from "../../../../src/useActionForm.ts";
 import { testApi as api } from "../../../apis.ts";
-import { StorybookErrorBoundary } from "../../StorybookErrorBoundary.tsx";
-import { makeAutocomponents } from "../../../../src/auto/shadcn/index.ts";
-import { elements } from "../index";
-import { makeShadcnAutoEnumInput } from "../../../../src/auto/shadcn/inputs/ShadcnAutoEnumInput.tsx";
+import { StorybookErrorBoundary } from "../../storybook/StorybookErrorBoundary.tsx";
+import { elements } from "../index.tsx";
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 
@@ -31,7 +31,7 @@ export default {
         <div style={{ width: "600px", backgroundColor: "white" }}>
           <Provider api={api}>
             <FormProvider {...useForm()}>
-              <StorybookErrorBoundary >
+              <StorybookErrorBoundary>
                 <elements.Card className="p-6 w-full bg-white shadow-lg rounded-lg">
                   <Story />
                 </elements.Card>
