+++ packages/react/spec/auto/storybook/StorybookErrorBoundary.tsx
