+++ packages/react/spec/auto/storybook/table/PolarisAutoTableActions.stories.jsx
@@ -1,15 +1,15 @@
-import { AppProvider, BlockStack, Box, LegacyCard } from "@shopify/polaris";
+import { AppProvider } from "@shopify/polaris";
 import translations from "@shopify/polaris/locales/en.json";
 import React from "react";
 import { Provider } from "../../../../src/GadgetProvider.tsx";
-import { PolarisAutoTable } from "../../../../src/auto/polaris/PolarisAutoTable.tsx";
 import { testApi as api } from "../../../apis.ts";
-import { StorybookErrorBoundary } from "../../StorybookErrorBoundary.tsx";
+import { StorybookErrorBoundary } from "../StorybookErrorBoundary.tsx";
+import { SelectableDesignSystemAutoTableStory } from "./SelectableDesignSystemAutoTableStory.tsx";
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 export default {
-  title: "Polaris/AutoTable/ActionParameters",
-  component: PolarisAutoTable,
+  title: "AutoTable/ActionParameters",
+  component: SelectableDesignSystemAutoTableStory,
   decorators: [
     // 👇 Defining the decorator in the preview file applies it to all stories
     (Story) => {
@@ -17,17 +17,9 @@ export default {
       return (
         <Provider api={api}>
           <AppProvider i18n={translations}>
-            <div style={{ width: "100%" }}>
-              <Box paddingBlockEnd="400">
-                <BlockStack gap="200">
-                  <LegacyCard>
-                    <StorybookErrorBoundary>
-                      <Story />
-                    </StorybookErrorBoundary>
-                  </LegacyCard>
-                </BlockStack>
-              </Box>
-            </div>
+            <StorybookErrorBoundary>
+              <Story />
+            </StorybookErrorBoundary>
           </AppProvider>
         </Provider>
       );
